Moved to subfolder 'strategies'
